initial port to 1.19

+++ src/main/java/com/enderio/core/api/client/gui/IGuiOverlay.java
@@ -19,9 +19,9 @@ public interface IGuiOverlay extends IHideable {
     //
     // boolean mouseMovedOrUp(int par1, int par2, int par3);
 
-    boolean handleMouseInput(int x, int y, int b);
+    boolean handleMouseInput(double x, double y, int b);
 
-    boolean isMouseInBounds(int mouseX, int mouseY);
+    boolean isMouseInBounds(double mouseX, double mouseY);
     
     void guiClosed();
